☀️ 53% – Declare nineteen more units; recover eight shared interfaces
Twenty-six units are now declared (up from seven): every composite that
compiles through the canonical route and keeps its exact members exact.
The scaffolder's include depth is fixed (three levels, not four — the
bug behind most of the initial 'collisions'), and true holes stay out
of the manifest since a listed retained member must include its draft.

The collision-resolution wave recovered eight shared declarations, each
proven by keeping every affected exact owner at zero differing
halfwords: Func_08077018 returns struct ItemDefinition *;
Data_03001f2c is the InventoryMenuState pointer; Func_080a4924 and
Func_080b6a60 take s32/u16* per their headers; Object_GetById is
s32(u32); RotateVectorByMagnitude and NormalizeVector take void *
vectors; Object_Destroy is void with an unprototyped parameter list;
Owner_GetState returns void * cast per caller; Func_08078ed8 returns
the Data_080844ec record.

Learned contracts, encoded in the workflow: wholly-exact units must
graduate as flat files under src/ with scheduler-style registration
(one member carries the path) — include-composites do not qualify, so
unit-080a448c's graduation is reverted and deferred; unit-08004838 and
unit-080a21b0 disagree with the ordered object manifest and stay
undeclared for individual debugging.

+++ games/gs1/src/inventory_menu/open_item_details.c
@@ -9,7 +9,7 @@ struct InventoryMenuDetailsState {
 
 s32 Func_080a10d0(s32 *, s32, s32, s32, s32, s32);
 void Func_080a22f4(void);
-void Func_080a4924(s32, u16);
+void Func_080a4924(s32, s32);
 #define InventoryMenu_DrawItemDetails Func_080a4924
 
 s32 InventoryMenu_OpenItemDetails(s32 item_index)
